Add endpoint for retrieving product listing
