+++ auth-api/src/auth_api/resources/__init__.py
@@ -40,6 +40,7 @@ from .user import API as USER_API
 from .bulk_user import API as BULK_USER_API
 from .user_settings import API as USER_SETTINGS_API
 from .org_products import API as ORG_PRODUCTS_API
+from .products import API as PRODUCTS_API
 from .notifications import API as NOTIFICATIONS_API
 
 
@@ -87,6 +88,7 @@ API.add_namespace(DOCUMENTS_API, path='/documents')
 API.add_namespace(CODES_API, path='/codes')
 API.add_namespace(ACCOUNTS_API, path='/accounts')
 API.add_namespace(ORG_PRODUCTS_API, path='/orgs/<string:org_id>/products')
+API.add_namespace(PRODUCTS_API, path='/products')
 API.add_namespace(NOTIFICATIONS_API, path='/users/<string:user_id>/org/<string:org_id>/notifications')
 API.add_namespace(USER_API, path='/users/<string:invitation_token>')
