+++ auth-api/src/auth_api/resources/products.py
@@ -13,17 +13,13 @@
 # limitations under the License.
 """API endpoints for managing an Org resource."""
 
-from flask import request
-from flask_restplus import Namespace, Resource, cors
+from flask_restplus import Namespace, Resource
 
 from auth_api import status as http_status
 from auth_api.exceptions import BusinessException
 from auth_api.jwt_wrapper import JWTWrapper
-from auth_api.schemas import ProductSubscriptionSchema
-from auth_api.schemas import utils as schema_utils
 from auth_api.services import Product as ProductService
 from auth_api.tracer import Tracer
-from auth_api.utils.roles import Role
 from auth_api.utils.util import cors_preflight
 
 API = Namespace('products', description='Endpoints for products management')
@@ -34,27 +30,13 @@ _JWT = JWTWrapper.get_instance()
 @cors_preflight('GET,POST,OPTIONS')
 @API.route('', methods=['GET', 'POST', 'OPTIONS'])
 class OrgProducts(Resource):
-    """Resource for managing product subscriptions."""
+    """Resource for managing products."""
 
     @staticmethod
-    @TRACER.trace()
-    @cors.crossdomain(origin='*')
-    @_JWT.has_one_of_roles([Role.STAFF_ADMIN.value])
-    def post(org_id):
-        """Post a new product subscription to the org using the request body."""
-        request_json = request.get_json()
-        valid_format, errors = schema_utils.validate(request_json, 'org_product_subscription')
-        if not valid_format:
-            return {'message': schema_utils.serialize(errors)}, http_status.HTTP_400_BAD_REQUEST
-
+    def get():
+        """Get a list of all products."""
         try:
-            subscriptions = ProductService.create_product_subscription(org_id, request_json)
-            if subscriptions is None:
-                response, status = {'message': 'Not authorized to perform this action'}, \
-                                   http_status.HTTP_401_UNAUTHORIZED
-            else:
-                response, status = {'subscriptions': ProductSubscriptionSchema().dump(subscriptions, many=True)}, \
-                                   http_status.HTTP_201_CREATED
+            response, status = ProductService.get_products(), http_status.HTTP_200_OK
         except BusinessException as exception:
             response, status = {'code': exception.code, 'message': exception.message}, exception.status_code
         return response, status
